@@ -1,12 +1,16 @@
 package de.bergtiger.claim.cmd;
 
+import net.md_5.bungee.api.chat.ClickEvent;
+import net.md_5.bungee.api.chat.TextComponent;
 import org.bukkit.Bukkit;
 import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
 
 import de.bergtiger.claim.Claims;
-import de.bergtiger.claim.data.Lang;
-import de.bergtiger.claim.data.Perm;
+import de.bergtiger.claim.data.language.Lang;
+import de.bergtiger.claim.data.permission.Perm;
+
+import java.util.ArrayList;
+import java.util.List;
 
 public class CmdHelp {
 	
@@ -15,61 +19,57 @@ public class CmdHelp {
 	}
 	
 	private void sendHelp(CommandSender cs) {
-		if (Perm.hasPermission(cs, Perm.CLAIM_ADMIN, Perm.CLAIM_CLAIM, Perm.CLAIM_PLUGIN, Perm.CLAIM_RELOAD)) {
-			if (cs instanceof Player) {
-				Player p = (Player) cs;
-				// Header
-				p.spigot().sendMessage(Lang.buildTC(Lang.CMD_HEADER.getValue()));
-				// Claim
-				if (Perm.hasPermission(cs, Perm.CLAIM_ADMIN, Perm.CLAIM_CLAIM))
-					p.spigot().sendMessage(
-							Lang.buildTC(Lang.CMD_INSERT.getValue(), "/claim " + Claim.CLAIM, Lang.CMD_HOVER_INSERT.getValue(), null));
-				// Info
-				if (Perm.hasPermission(cs, Perm.CLAIM_ADMIN, Perm.CLAIM_INFO))
-					p.spigot().sendMessage(
-							Lang.buildTC(Lang.CMD_INFO.getValue(), "/claim " + Claim.INFO, Lang.CMD_HOVER_INFO.getValue(), null));
-				// List
-				if (Perm.hasPermission(cs, Perm.CLAIM_ADMIN, Perm.CLAIM_LIST))
-					p.spigot().sendMessage(
-							Lang.buildTC(Lang.CMD_LIST.getValue(), "/claim " + Claim.LIST, Lang.CMD_HOVER_LIST.getValue(), null));
-				// Set
-				if (Perm.hasPermission(cs, Perm.CLAIM_ADMIN, Perm.CLAIM_SET))
-					p.spigot().sendMessage(
-							Lang.buildTC(Lang.CMD_SET.getValue(), null, Lang.CMD_HOVER_SET.getValue(), "/claim " + Claim.SET + " "));
-				// Plugin
-				if (Perm.hasPermission(cs, Perm.CLAIM_ADMIN, Perm.CLAIM_PLUGIN))
-					p.spigot().sendMessage(
-							Lang.buildTC(Lang.CMD_PLUGIN.getValue(), "/claim " + Claim.PLUGIN, Lang.CMD_HOVER_PLUGIN.getValue(), null));
-				// Reload
-				if (Perm.hasPermission(cs, Perm.CLAIM_ADMIN, Perm.CLAIM_RELOAD))
-					p.spigot().sendMessage(
-							Lang.buildTC(Lang.CMD_RELOAD.getValue(), "/claim " + Claim.RELOAD, Lang.CMD_HOVER_RELOAD.getValue(), null));
-				// Footer
-				p.spigot().sendMessage(Lang.buildTC(Lang.CMD_FOOTER.getValue()));
-			} else {
-				// Header
-				cs.sendMessage(Lang.CMD_HEADER.get());
-				// Claim
-				if (Perm.hasPermission(cs, Perm.CLAIM_ADMIN, Perm.CLAIM_CLAIM))
-					cs.sendMessage(Lang.CMD_INSERT.get());
-				// Set
-				if (Perm.hasPermission(cs, Perm.CLAIM_ADMIN, Perm.CLAIM_SET))
-					cs.sendMessage(Lang.CMD_SET.get());
-				// Info
-				if (Perm.hasPermission(cs, Perm.CLAIM_ADMIN, Perm.CLAIM_INFO))
-					cs.sendMessage(Lang.CMD_INFO.get());
-				// List
-				if (Perm.hasPermission(cs, Perm.CLAIM_ADMIN, Perm.CLAIM_LIST))
-					cs.sendMessage(Lang.CMD_LIST.get());
-				// Plugin
-				if (Perm.hasPermission(cs, Perm.CLAIM_ADMIN, Perm.CLAIM_PLUGIN))
-					cs.sendMessage(Lang.CMD_PLUGIN.get());
-				// Reload
-				if (Perm.hasPermission(cs, Perm.CLAIM_ADMIN, Perm.CLAIM_RELOAD))
-					cs.sendMessage(Lang.CMD_RELOAD.get());
-				// Footer
-				cs.sendMessage(Lang.CMD_FOOTER.get());
+		if (Perm.hasPermissionAny(cs)) {
+			List<TextComponent> components = new ArrayList<>();
+			// Header
+			components.add(Lang.build(Lang.CMD_HEADER));
+			components.add(Lang.newLine());
+			// Claim
+			if (Perm.hasPermission(cs, Perm.CLAIM_ADMIN, Perm.CLAIM_CLAIM)) {
+				components.add(Lang.build(Lang.CMD_INSERT, String.format("/claim %s", Claim.CLAIM), Lang.build(Lang.CMD_HOVER_INSERT), null));
+				components.add(Lang.newLine());
 			}
+			// Info
+			if (Perm.hasPermission(cs, Perm.CLAIM_ADMIN, Perm.CLAIM_INFO)) {
+				components.add(
+						Lang.build(Lang.CMD_INFO, String.format("/claim %s", Claim.INFO), Lang.build(Lang.CMD_HOVER_INFO), null));
+				components.add(Lang.newLine());
+			}
+			// List
+			if (Perm.hasPermission(cs, Perm.CLAIM_ADMIN, Perm.CLAIM_LIST)) {
+				components.add(
+						Lang.build(Lang.CMD_LIST, String.format("/claim %s", Claim.LIST), Lang.build(Lang.CMD_HOVER_LIST), null));
+				components.add(Lang.newLine());
+			}
+			// Set
+			if (Perm.hasPermission(cs, Perm.CLAIM_ADMIN, Perm.CLAIM_SET)) {
+				components.add(
+						Lang.build(Lang.CMD_SET, null, Lang.build(Lang.CMD_HOVER_SET), String.format("/claim %s ", Claim.SET)));
+				components.add(Lang.newLine());
+			}
+			// Plugin
+			if (Perm.hasPermission(cs, Perm.CLAIM_ADMIN, Perm.CLAIM_PLUGIN)) {
+				components.add(
+						Lang.build(Lang.CMD_PLUGIN, String.format("/claim %s", Claim.PLUGIN), Lang.build(Lang.CMD_HOVER_PLUGIN), null));
+				components.add(Lang.newLine());
+			}
+			// Reload
+			if (Perm.hasPermission(cs, Perm.CLAIM_ADMIN, Perm.CLAIM_RELOAD)) {
+				components.add(
+						Lang.build(Lang.CMD_RELOAD, String.format("/claim %s", Claim.RELOAD), Lang.build(Lang.CMD_HOVER_RELOAD), null));
+				components.add(Lang.newLine());
+			}
+			// Wiki
+			if (Perm.hasPermission(cs, Perm.CLAIM_ADMIN)) {
+				TextComponent tc = Lang.build(Lang.CMD_WIKI, null, Lang.build(Lang.CMD_HOVER_WIKI), null);
+				tc.setClickEvent(new ClickEvent(ClickEvent.Action.OPEN_URL, "https://github.com/Bergtiger/TigerClaim/wiki/TigerClaim"));
+				components.add(tc);
+				components.add(Lang.newLine());
+			}
+			// Footer
+			components.add(Lang.build(Lang.CMD_FOOTER));
+			// show message
+			cs.spigot().sendMessage(Lang.combine(components));
 		} else {
 			cs.sendMessage(Lang.NOPERMISSION.get());
 		}
