@@ -1,16 +1,19 @@
 package de.bergtiger.claim.cmd;
 
-import static de.bergtiger.claim.data.Cons.FLAG;
-import static de.bergtiger.claim.data.Cons.VALUE;
+import static de.bergtiger.claim.data.configuration.Config.*;
+import static de.bergtiger.claim.data.language.Lang.*;
 
+import net.md_5.bungee.api.chat.TextComponent;
 import org.bukkit.Bukkit;
 import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
 
 import de.bergtiger.claim.Claims;
-import de.bergtiger.claim.data.Config;
-import de.bergtiger.claim.data.Lang;
-import de.bergtiger.claim.data.Perm;
+import de.bergtiger.claim.data.configuration.Config;
+import de.bergtiger.claim.data.language.Lang;
+import de.bergtiger.claim.data.permission.Perm;
+
+import java.util.ArrayList;
+import java.util.List;
 
 public class CmdPlugin {
 
@@ -19,121 +22,105 @@ public class CmdPlugin {
 	}
 	
 	private void sendPluginInfo(CommandSender cs) {
+		String cmd = "/claim set %s ";
 		if (Perm.hasPermission(cs, Perm.CLAIM_ADMIN, Perm.CLAIM_PLUGIN)) {
-			Config c = Config.inst();
-			if (cs instanceof Player) {
-				Player p = (Player) cs;
-				// Header
-				p.spigot().sendMessage(Lang.buildTC(Lang.PLUGIN_HEADER.getValue()));
-				// Version
-				p.spigot().sendMessage(Lang.buildTC(
-						Lang.PLUGIN_VERSION.getValue().replace(VALUE, Claims.inst().getDescription().getVersion())));
-				// Page Length
-				p.spigot()
-						.sendMessage(Lang.buildTC(
-								Lang.PLUGIN_PAGE_LENGTH.getValue().replace(VALUE, 
-										c.hasValue(Config.PAGE_LENGTH) ? c.getValue(Config.PAGE_LENGTH).toString()
-												: "-"),
-								null, Lang.PLUGIN_HOVER_PAGE_LENGTH.getValue(), "/claim set " + CmdSet.CMD_PAGE_LENGTH + " "));
-				// Radius
-				p.spigot()
-						.sendMessage(Lang.buildTC(
-								Lang.PLUGIN_RADIUS.getValue().replace(VALUE,
-										c.hasValue(Config.REGION_RADIUS) ? c.getValue(Config.REGION_RADIUS).toString()
-												: "-"),
-								null, Lang.PLUGIN_HOVER_RADIUS.getValue(), "/claim set " + CmdSet.CMD_RADIUS + " "));
-				// Gap
-				p.spigot()
-						.sendMessage(
-								Lang.buildTC(
-										Lang.PLUGIN_GAP.getValue().replace(VALUE,
-												c.hasValue(Config.REGION_GAP) ? c.getValue(Config.REGION_GAP).toString()
-														: "-"),
-										null, Lang.PLUGIN_HOVER_GAP.getValue(), "/claim set " + CmdSet.CMD_GAP + " "));
-				// Expand Vert
-				p.spigot().sendMessage(Lang.buildTC(
-						Lang.PLUGIN_EXPANDVERT.getValue().replace(VALUE,
-								c.hasValue(Config.REGION_EXPAND_VERT) ? c.getValue(Config.REGION_EXPAND_VERT).toString()
-										: "-"),
-						null, Lang.PLUGIN_HOVER_EXPANDVERT.getValue(),
-						"/claim set " + CmdSet.CMD_EXPAND_VERT + " "
-								+ (c.hasValue(Config.REGION_EXPAND_VERT)
-										? !Boolean.valueOf(c.getValue(Config.REGION_EXPAND_VERT).toString())
-										: "")));
-				// Overlapping
-				p.spigot().sendMessage(Lang.buildTC(
-						Lang.PLUGIN_OVERLAPPING.getValue().replace(VALUE,
-								c.hasValue(Config.REGION_OVERLAPPING) ? c.getValue(Config.REGION_OVERLAPPING).toString()
-										: "-"),
-						null, Lang.PLUGIN_HOVER_OVERLAPPING.getValue(),
-						"/claim set " + CmdSet.CMD_OVERLAPPING + " "
-								+ (c.hasValue(Config.REGION_OVERLAPPING)
-										? !Boolean.valueOf(c.getValue(Config.REGION_OVERLAPPING).toString())
-										: "")));
-				// Pattern
-				p.spigot()
-						.sendMessage(Lang.buildTC(
-								Lang.PLUGIN_PATTERN_ID.getValue().replace(VALUE,
-										c.hasValue(Config.REGION_PATTERN) ? c.getValue(Config.REGION_PATTERN).toString()
-												: "-"),
-								null, Lang.PLUGIN_HOVER_PATTERN_ID.getValue(), "/claim set " + CmdSet.CMD_PATTERN + " "));
-				// Time Pattern
-				p.spigot()
-						.sendMessage(Lang.buildTC(
-								Lang.PLUGIN_PATTERN_TIME.getValue().replace(VALUE,
-										c.hasValue(Config.TIME_PATTERN) ? c.getValue(Config.TIME_PATTERN).toString()
-												: "-"),
-								null, Lang.PLUGIN_HOVER_PATTERN_TIME.getValue(), "/claim set " + CmdSet.CMD_TIME + " "));
-				// Flags
-				p.spigot().sendMessage(Lang.buildTC(Lang.PLUGIN_FLAGS.getValue(), null, Lang.PLUGIN_HOVER_FLAGS.get(),
-						"/claim set flags "));
-				if (c.hasFlags()) {
-					c.getFlags().forEach((f, v) -> {
-						p.spigot()
-								.sendMessage(Lang.buildTC(
-										Lang.PLUGIN_FLAG_LIST.get().replace(FLAG, f.getName()).replace(VALUE,
-												v.toString()),
-										null, null, "/claim set " + CmdSet.CMD_FLAG + " " + f.getName()));
-					});
-				}
-				// Footer
-				p.spigot().sendMessage(Lang.buildTC(Lang.PLUGIN_FOOTER.get()));
-			} else {
-				// Header
-				cs.sendMessage(Lang.PLUGIN_HEADER.get());
-				// Version
-				cs.sendMessage(Lang.PLUGIN_VERSION.get().replace(VALUE, Claims.inst().getDescription().getVersion()));
-				// Page Length
-				cs.sendMessage(Lang.PLUGIN_PAGE_LENGTH.get().replace(VALUE,
-						c.hasValue(Config.PAGE_LENGTH) ? c.getValue(Config.PAGE_LENGTH).toString() : "-"));
-				// Radius
-				cs.sendMessage(Lang.PLUGIN_RADIUS.get().replace(VALUE,
-						c.hasValue(Config.REGION_RADIUS) ? c.getValue(Config.REGION_RADIUS).toString() : "-"));
-				// Gap
-				cs.sendMessage(Lang.PLUGIN_GAP.get().replace(VALUE,
-						c.hasValue(Config.REGION_GAP) ? c.getValue(Config.REGION_GAP).toString() : "-"));
-				// Expand Vert
-				cs.sendMessage(Lang.PLUGIN_EXPANDVERT.get().replace(VALUE,
-						c.hasValue(Config.REGION_EXPAND_VERT) ? c.getValue(Config.REGION_EXPAND_VERT).toString()
-								: "-"));
-				// Pattern
-				cs.sendMessage(Lang.PLUGIN_PATTERN_ID.get().replace(VALUE,
-						c.hasValue(Config.REGION_PATTERN) ? c.getValue(Config.REGION_PATTERN).toString() : "-"));
-				// Time Pattern
-				cs.sendMessage(Lang.PLUGIN_PATTERN_TIME.get().replace(VALUE,
-						c.hasValue(Config.TIME_PATTERN) ? c.getValue(Config.TIME_PATTERN).toString() : "-"));
-				// Flags
-				cs.sendMessage(Lang.PLUGIN_FLAGS.get());
-				if (c.hasFlags()) {
-					c.getFlags().forEach((f, v) -> {
-						if ((f != null) && (v != null))
-							cs.sendMessage(Lang.PLUGIN_FLAG_LIST.get().replace(FLAG, f.getName()).replace(VALUE,
-									v.toString()));
-					});
-				}
-				// Footer
-				cs.sendMessage(Lang.PLUGIN_FOOTER.get());
+			List<TextComponent> components = new ArrayList<>();
+			// Header
+			components.add(build(PLUGIN_HEADER));
+			components.add(newLine());
+			// Version
+			components.add(build(
+					PLUGIN_VERSION.replace(VALUE, Claims.inst().getDescription().getVersion())));
+			components.add(newLine());
+			// Page Length
+			components.add(build(
+					PLUGIN_PAGE_LENGTH.replace(VALUE, Integer.toString(getInt(PAGE_LENGTH))),
+					null,
+					build(PLUGIN_HOVER_PAGE_LENGTH),
+					String.format(cmd, CmdSet.CMD_PAGE_LENGTH)));
+			components.add(newLine());
+			// Radius
+			components.add(build(
+					PLUGIN_RADIUS.replace(VALUE, Integer.toString(getInt(REGION_RADIUS))),
+					null,
+					build(PLUGIN_HOVER_RADIUS),
+					String.format(cmd, CmdSet.CMD_RADIUS)));
+			components.add(newLine());
+			// Gap
+			components.add(build(
+					PLUGIN_GAP.replace(VALUE, Integer.toString(getInt(REGION_GAP))),
+					null,
+					build(PLUGIN_HOVER_GAP),
+					String.format(cmd, CmdSet.CMD_GAP)));
+			components.add(newLine());
+			// Expand Vert
+			components.add(build(
+					PLUGIN_EXPANDVERT.replace(VALUE, Boolean.toString(getBoolean(REGION_EXPAND_VERT))),
+					null,
+					build(PLUGIN_HOVER_EXPANDVERT),
+					String.format(cmd, CmdSet.CMD_EXPAND_VERT)
+							+ (hasValue(Config.REGION_EXPAND_VERT)
+									? Boolean.toString(!getBoolean(Config.REGION_EXPAND_VERT)) : "")));
+			components.add(newLine());
+			// Overlapping
+			components.add(build(
+					PLUGIN_OVERLAPPING.replace(VALUE, Boolean.toString(getBoolean(REGION_OVERLAPPING))),
+					null,
+					build(PLUGIN_HOVER_OVERLAPPING),
+					"/claim set " + CmdSet.CMD_OVERLAPPING + " "
+							+ (hasValue(Config.REGION_OVERLAPPING)
+									? Boolean.toString(!getBoolean(REGION_OVERLAPPING)) : "")));
+			components.add(newLine());
+			// Pattern
+			components.add(build(
+					PLUGIN_PATTERN_ID.replace(VALUE, hasValue(REGION_PATTERN) ? getString(REGION_PATTERN) : "-"),
+					null,
+					build(PLUGIN_HOVER_PATTERN_ID),
+					String.format(cmd, CmdSet.CMD_PATTERN)));
+			components.add(newLine());
+			// Time Pattern
+			components.add(build(
+					PLUGIN_PATTERN_TIME.replace(VALUE, hasValue(TIME_PATTERN) ? getString(TIME_PATTERN) : "-"),
+					null,
+					build(PLUGIN_HOVER_PATTERN_TIME),
+					String.format(cmd, CmdSet.CMD_TIME)));
+			components.add(newLine());
+			// height
+			// min
+			components.add(build(
+					PLUGIN_HEIGHT_MIN.replace(VALUE, Integer.toString(getInt(REGION_MIN_HEIGHT))),
+					null,
+					build(PLUGIN_HOVER_HEIGHT_MIN),
+					String.format(cmd, CmdSet.CMD_HEIGHT_MIN)));
+			components.add(newLine());
+			// max
+			components.add(build(
+					PLUGIN_HEIGHT_MAX.replace(VALUE, Integer.toString(getInt(REGION_MAX_HEIGHT))),
+					null,
+					build(PLUGIN_HOVER_HEIGHT_MAX),
+					String.format(cmd, CmdSet.CMD_HEIGHT_MAX)));
+			components.add(newLine());
+			// Flags
+			components.add(build(
+					PLUGIN_FLAGS,
+					null,
+					build(PLUGIN_HOVER_FLAGS),
+					String.format(cmd, CmdSet.CMD_FLAG)));
+			components.add(newLine());
+			if (hasValue(REGION_FLAGS)) {
+				getFlags().forEach((f, v) -> {
+					components.add(build(
+							PLUGIN_FLAG_LIST.replace(FLAG, f.getName()).replace(VALUE, v.toString()),
+							null,
+							null,
+							String.format("/claim %s %s %s ", Claim.SET, CmdSet.CMD_FLAG, f.getName())));
+					components.add(newLine());
+				});
 			}
+			// Footer
+			components.add(build(Lang.PLUGIN_FOOTER));
+			// show message
+			cs.spigot().sendMessage(Lang.combine(components));
 		} else {
 			cs.sendMessage(Lang.NOPERMISSION.get());
 		}
